Extract sources file
